+++ CHANGELOG.md
@@ -86,3 +86,4 @@ Semua perubahan yang dilakukan akan didokumentasikan di dalam file ini.
 - Menambahkan **pagination** ke `rooms`.
 - Menambahkan **pagination** ke `dashboard`.
 - Menambahkan **pagination** ke `bookings`.
+- Menambahkan **footer** ke `booking.html`, `booking-form.html` dan `admin-approval.html`,